SKALE-3835 Fix server test
