+++ tests/server_test.py
@@ -91,10 +91,7 @@ def test_send_transaction_errored(skale_bp):
         'transaction_dict': tx_dict_str
     })
     assert response['data'] is None
-    assert response['error'] in (
-        "Transaction must include these fields: {'gas', 'gasPrice'}",
-        "Transaction must include these fields: {'gasPrice', 'gas'}"
-    )
+    assert response['error'] == "Transaction must include these fields: {'gasPrice'}"
 
 
 def test_sign_hash(skale_bp):
